requirements.txt
open3d==0.19.0
termcolor==3.1.0
tqdm==4.67.1
einops==0.8.1
transforms3d==0.4.2
msgpack-numpy==0.4.8
lmdb==1.6.2
h5py==3.14.0
hydra-core==0.11.3
pytorch-lightning==2.5.2
scikit-image==0.24.0
black==25.1.0
usort==1.0.8.post1
flake8==7.3.0
matplotlib==3.9.4
jupyter==1.1.1
imageio==2.37.0
fvcore==0.1.5.post20221221
plotly==6.2.0
opencv-python==4.11.0.86
markdown==3.8.2
rich==15.0.0
tensorboard==2.19.0
ninja==1.13.0
wandb==0.20.1